rdbms: Bump ChronologyProtector cache key version
In I78096d893a6a (96af469d8) we moved from v2 to v3, which uses
a plain array instead of an object in serialised PHP.

In Idd79aeb630bf (4e4a02d2d00) I simplified the code thinking the
data shape remains the same. While this is true for what we send
and receive, it ends up differently internally.

Bug: T317606
Change-Id: I9cda0750f6d6662b7274808fb9b6de56a4edc2ff