Check whether title actually exists
Code a couple of lines below is dependant on $title
existing and having an id, but that's not explicitly
being checked.

If the result is not a valid title, there isn't going
to be a thumbnail for it.

Bug: T319798
Change-Id: I22c2ab886b3a1b1fdf2df062ffa76cd00d4d25fd
(cherry picked from commit 58d9b4f8f06b88f62d169f0a922f4f51d19bd4eb)