Fix LanguageVariantConverter test
Accept sr-Latn as well as sr-el as the language code for Serbian with
latin script.

This was broken when the parsoid library started to use BCP-47 codes
rather than internal MediaWiki codes. For now, we accept both, so we are
compatible with the version of the parsoid lib currently in the vendor
repo as well as the version picked by composer update.

Bug: T323985
Change-Id: If0b02be4f391b31fb75e2ad51e199a83707b0e3c