Parsoid: don't bypass ParserCache when using Title
We were bypassing the ParserCache when the page was given as a Title
instead of a RevisionRecord, because
ParsoidOutputAccess::getParserOutput used to require a RevisionRecord.
This is no longer the case (per a02be0b3f85c7a3b75ece4806338cea63bd686e5).

This fixes a performance issue with VisuaEditor: since VE passes a Title
object to HtmlOutputRendererHelper, we ended up not using the parser
cache when loading VE. This caused load times to go up by more than a
second on average on the wiki where we enabled direct mode for VE
(hewiki).

Change-Id: I8d483971fd116e74e5e0923afce0abfd1402ae10