linker: Check for null parser in Linker::makeThumbLink2
Just as fix for the null access.
Adding a module in the middle of Linker could be wrong as that function
is used on many places in core. Not sure if other places needs the
module as well.

Bug: T337794
Change-Id: I360b1ded9ef14b032bcf91c1699bcc325441bd7d
(cherry picked from commit 9e53a1a1ddb2512d09daed06d0e8c60d3d666049)