file: Make pre-gen rendering of multi-page files (pdf, ...) serial
Instead of unleashing 200 jobs to production that would hit thumbor
around the same time, just push the first page, get it rendered, and
push the next job until the limit is reached.

This is not to reduce pressure on jobqueue, it's to reduce pressue on
thumbor expensive jobs.

This is similar to what refreshlinks does, after editing on a heavily
used template, it reparses 100 usages, queues the next 100 batch, and
so on.

Bug: T337649
Change-Id: Iad6dd8828c79e3363f7256f2395e95046122530e
(cherry picked from commit 131d3abd2cf8117d1a807040a7cb26f27f674f3f)